perf: lazy-load heavy deps and add Lighthouse CI budgets
- pdfjs-dist loaded at runtime inside BookReaderClient effect
  (worker ?url import preserved); library list pages no longer pay for it
- recharts split out of dashboard home + earnings initial chunks
  via next/dynamic ssr:false children with skeleton fallbacks
- Jitsi meeting client dynamic ssr:false, loads only on join
- optimizePackageImports for lucide-react/react-icons
- @next/bundle-analyzer behind ANALYZE=true + npm run analyze
- Lighthouse CI job (/, /login) with committed lighthouserc.json
  budgets; reuses build stub env block
