+++ app/[locale]/dashboard/earnings/page.jsx
@@ -1,5 +1,6 @@
 "use client";
 import { useState } from "react";
+import dynamic from "next/dynamic";
 import {
   DollarSign,
   ShoppingCart,
@@ -12,12 +13,6 @@ import { Button } from "@/components/ui/button";
 import { PageShell } from "@/components/ui/page-shell";
 import { PageHeader } from "@/components/ui/page-header";
 import { EmptyState } from "@/components/ui/empty-state";
-import {
-  ChartContainer,
-  ChartTooltip,
-  ChartTooltipContent,
-} from "@/components/ui/chart";
-import { BarChart, Bar, XAxis, YAxis, CartesianGrid } from "recharts";
 import WalletConnectButton from "@/components/stellar/WalletConnectButton";
 import useEarnings from "@/hooks/useEarnings";
 import Link from "next/link";
@@ -31,10 +26,15 @@ import {
 const statusStyles = {
   confirmed: "bg-secondary/10 text-secondary border-secondary/20",
   pending: "bg-amber-100 text-amber-700 border-amber-200",
-  submitted: "bg-sky-100 text-sky-700 border-sky-200",
+  submitted: "bg-sky-100 text-sky-200 border-sky-200",
   failed: "bg-red-100 text-red-600 border-red-200",
   expired: "bg-ink/5 text-ink-muted border-accent/10",
 };
+
+const RevenueChart = dynamic(() => import("./RevenueChart"), {
+  ssr: false,
+  loading: () => <Skeleton className="h-72 w-full rounded-xl" />,
+});
 
 /* ── building blocks (design-system consistent) ── */
 
@@ -151,7 +151,6 @@ export default function EarningsPage() {
 
   const [chartRange, setChartRange] = useState("30d");
   const chartData = revenueChartData(chartRange);
-
   if (!hasWallet && !isLoading) {
     return (
       <PageShell>
@@ -196,10 +195,6 @@ export default function EarningsPage() {
       </PageShell>
     );
   }
-
-  const chartConfig = {
-    revenue: { label: "Revenue", color: "#009900" },
-  };
 
   return (
     <PageShell>
@@ -302,48 +297,7 @@ export default function EarningsPage() {
                 No revenue data for this period
               </div>
             ) : (
-              <ChartContainer config={chartConfig} className="h-72 w-full">
-                <BarChart data={chartData}>
-                  <CartesianGrid
-                    strokeDasharray="3 3"
-                    vertical={false}
-                    stroke="#265902"
-                    strokeOpacity={0.12}
-                  />
-                  <XAxis
-                    dataKey="date"
-                    tickLine={false}
-                    axisLine={false}
-                    tickMargin={8}
-                    tickFormatter={(val) => {
-                      const d = new Date(val + "T00:00:00");
-                      return d.toLocaleDateString("en-US", {
-                        month: "short",
-                        day: "numeric",
-                      });
-                    }}
-                  />
-                  <YAxis
-                    tickLine={false}
-                    axisLine={false}
-                    tickMargin={8}
-                    tickFormatter={(val) => `$${val}`}
-                  />
-                  <ChartTooltip
-                    cursor={{ fill: "rgba(0,153,0,0.06)" }}
-                    content={
-                      <ChartTooltipContent
-                        formatter={(value) => `$${value.toFixed(2)}`}
-                      />
-                    }
-                  />
-                  <Bar
-                    dataKey="revenue"
-                    fill="var(--color-revenue)"
-                    radius={[6, 6, 0, 0]}
-                  />
-                </BarChart>
-              </ChartContainer>
+              <RevenueChart data={chartData} />
             )}
           </Panel>
